refactor: restructure activities module by moving controller and service files, and update imports in app module
