+++ backend/app/src/app.module.ts
@@ -25,7 +25,7 @@ import { ActivitiesModule } from './module/activities/activities.module';
 import { EvoucherModule } from './module/evoucher/evoucher.module';
 import { EvoucherTypeModule } from './module/evoucher/evoucher-type.module';
 import { EvoucherCodeModule } from './module/evoucher/evoucher-code.module';
-import { ActivitiesTypeModule } from './module/activities-type/activities-type.module';
+import { ActivitiesTypeModule } from './module/activities/activities-type.module';
 import { NotificationsModule } from './module/notifications/notifications.module';
 
 @Module({
