+++ backend/app/src/module/users/users.controller.ts
@@ -10,6 +10,7 @@ import {
   Patch,
   Post,
   Req,
+  UnauthorizedException,
 } from '@nestjs/common';
 
 import { UsersService } from './users.service';
@@ -22,11 +23,16 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { AutoCacheInterceptor } from 'src/pkg/cache/auto-cache.interceptor';
 import { FastifyRequest } from 'fastify';
 import { UserUploadDirectDto } from './dto/upload.user.dto';
+import { ActivitiesService } from '../activities/service/activities.service';
+
 @UseGuards(PermissionsGuard)
 @UseInterceptors(AutoCacheInterceptor)
 @Controller('users')
 export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(
+    private readonly usersService: UsersService,
+    private readonly activitiesService: ActivitiesService,
+  ) {}
 
   @Post()
   @Public()
@@ -47,13 +53,6 @@ export class UsersController {
   @CacheKey('users')
   async findAll(@Query() query: Record<string, string>) {
     return this.usersService.findAll(query);
-  }
-
-  @Get('by-query')
-  @Permissions('users:read')
-  @CacheKey('users:by-query')
-  async findAllByQuery(@Query() query: Record<string, string>) {
-    return this.usersService.findAllByQuery(query);
   }
 
   @Get('statistics')
@@ -86,6 +85,21 @@ export class UsersController {
     return this.usersService.findOneByQuery({
       _id: userId,
     });
+  }
+
+  @Get('activities')
+  @CacheKey('users:activities:$req.user')
+  getUserActivities(
+    @Req() req: FastifyRequest & { user?: { _id?: string; id?: string } },
+  ) {
+    const user = req.user;
+    const userId: string | undefined = user?._id ?? user?.id;
+
+    if (!userId) {
+      throw new UnauthorizedException('User not found');
+    }
+
+    return this.activitiesService.findActivitiesByUserId(userId);
   }
 
   @Patch(':id')
