+++ backend/app/src/module/users/users.module.ts
@@ -4,8 +4,9 @@ import { UsersController } from './users.controller';
 import { MongooseModule } from '@nestjs/mongoose';
 import { User, UserSchema } from './schemas/user.schema';
 import { Major, MajorSchema } from '../majors/schemas/major.schema';
-import { Role, RoleSchema } from '../role/schemas/role.schema';
-import { UserInitializerService } from './users.initializer.service';
+import { Role, RoleSchema } from '../role/schemas/role.schema';3
+import { Activities, ActivitiesSchema } from '../activities/schema/activities.schema';
+import { ActivitiesService } from '../activities/service/activities.service';
 
 @Module({
   imports: [
@@ -13,10 +14,11 @@ import { UserInitializerService } from './users.initializer.service';
       { name: User.name, schema: UserSchema },
       { name: Role.name, schema: RoleSchema },
       { name: Major.name, schema: MajorSchema },
+      { name: Activities.name, schema: ActivitiesSchema },
     ]),
   ],
   controllers: [UsersController],
-  providers: [UsersService, UserInitializerService],
-  exports: [UsersService],
+  providers: [UsersService, ActivitiesService],
+  exports: [UsersService, MongooseModule],
 })
 export class UsersModule {}
